// src/models/_foodb-models/db.d.ts
export interface nutrientsAttribute {
	id: number;
	legacyId?: number;
	type: string;
	publicId: string;
	name: string;
	export?: number;
	state?: string;
	annotationQuality?: string;
	description?: string;
	wikipediaId?: string;
	comments?: string;
	dfcId?: string;
	dukeId?: string;
	eafusId?: string;
	dfcName?: string;
	compoundSource?: string;
	metabolism?: string;
	synthesisCitations?: string;
	generalCitations?: string;
	creatorId?: number;
	updaterId?: number;
	createdAt?: Date;
	updatedAt?: Date;
}

export interface nutrientsInstance extends Sequelize.Instance<nutrientsAttribute>, nutrientsAttribute {
}